style: OpenAiClient 로 네이밍 수정 - #28
